+++ Sources/Samples/ShuttlePB/ShuttlePB.h
@@ -18,7 +18,7 @@
 // Shuttle-PB class interface
 // ==============================================================
 
-class ShuttlePB : public VESSEL4 
+class ShuttlePB : public VESSEL4
 {
 public:
 	ShuttlePB(OBJHANDLE hVessel, int flightmodel);
@@ -39,10 +39,10 @@ private:
 	UACS::API::VslAstrInfo astrInfo;
 	UACS::API::VslCargoInfo cargoInfo;
 
-	int cargoIndex = 0;  // For the cargo selection on the HUD
+	size_t cargoIdx{};  // For the cargo selection on the HUD
 	char buffer[256];    // To draw on the HUD
-	const char* message; // To show messages on the HUD
-	double timer = 0;    // To show the messages for 5 seconds on the HUD
+	const char* message{}; // To show messages on the HUD
+	double timer{ 5 };    // To show the messages for 5 seconds on the HUD
 
 	void SetStatusLanded();
 
@@ -51,6 +51,30 @@ private:
 	static void hlift(VESSEL* v, double aoa, double M, double Re,
 		void* context, double* cl, double* cm, double* cd);
 };
+
+inline MATRIX3 RotationMatrix(VECTOR3 angles)
+{
+	const MATRIX3 RM_X = _M(1, 0, 0, 0, cos(angles.x), -sin(angles.x), 0, sin(angles.x), cos(angles.x));
+	const MATRIX3 RM_Y = _M(cos(angles.y), 0, sin(angles.y), 0, 1, 0, -sin(angles.y), 0, cos(angles.y));
+	const MATRIX3 RM_Z = _M(cos(angles.z), -sin(angles.z), 0, sin(angles.z), cos(angles.z), 0, 0, 0, 1);
+
+	return mul(RM_X, mul(RM_Y, RM_Z));
+}
+
+inline void SetGroundRotation(VESSELSTATUS2& status, double height)
+{
+	const MATRIX3 rot1 = RotationMatrix({ 0, PI05 - status.surf_lng, 0 });
+	const MATRIX3 rot2 = RotationMatrix({ -status.surf_lat, 0, 0 });
+	const MATRIX3 rot3 = RotationMatrix({ 0, 0, PI + status.surf_hdg });
+	const MATRIX3 rot4 = RotationMatrix({ PI05, 0, 0 });
+	const MATRIX3 RotMatrix_Def = mul(rot1, mul(rot2, mul(rot3, rot4)));
+
+	status.arot.x = atan2(RotMatrix_Def.m23, RotMatrix_Def.m33);
+	status.arot.y = -asin(RotMatrix_Def.m13);
+	status.arot.z = atan2(RotMatrix_Def.m12, RotMatrix_Def.m11);
+
+	status.vrot.x = height;
+}
 
 // ==============================================================
 // Some vessel parameters